+++ src/main/java/me/ajlane/geo/repast/fire/DefaultFireManager.java
@@ -64,7 +64,7 @@ public class DefaultFireManager implements FireManager {
         break;
       }
       // firesStarted.add(initialIgnitionPoint);
-      this.fireSpreader.spreadFire(initialIgnitionPoint, this.vegetationMoistureParam);
+      this.fireSpreader.spreadFire(initialIgnitionPoint);
     }
   }