+++ src/main/java/me/ajlane/geo/repast/fire/FireSpreader.java
@@ -11,9 +11,7 @@ public interface FireSpreader {
    * inspiration behind the fire spread mechanism.
    *
    * @param ignitionPoint Point in the landscape where the fire starts
-   * @param fuelMoistureFactor Dimensionless factor parameterising the amount of moisture in the
-   *        fuel at the time of the fire
    */
-  void spreadFire(GridPoint ignitionPoint, double fuelMoistureFactor);
+  void spreadFire(GridPoint ignitionPoint);
 
 }